DASH-310 Move reused ContentContainer, MainHeaderContainer, MainLayoutContainer to common components folder
