+++ client/src/components/MainHeaderContainer/index.ts
@@ -1,0 +1,1 @@
+export { MainHeaderContainer } from './MainHeaderContainer';
