+++ client/src/components/MainLayoutContainer/MainLayoutContainer.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../assets/styles/colors';
+@import '../../assets/styles/colors';
 
 .root {
   min-height: 100vh;
